revert to d4528e6079d91aeb9c1991d3f65493bb535e6382

+++ pandas/pandas/tests/series/methods/test_asof.py
@@ -63,8 +63,7 @@ class TestSeriesAsof:
 
         N = 30
         rng = date_range("1/1/1990", periods=N, freq="53s")
-        # Explicit cast to float avoid implicit cast when setting nan
-        ts = Series(np.arange(N), index=rng, dtype="float")
+        ts = Series(np.arange(N), index=rng)
         ts.iloc[5:10] = np.NaN
         ts.iloc[15:20] = np.NaN
