+++ pandas/pandas/tests/series/methods/test_interpolate.py
@@ -703,8 +703,7 @@ class TestSeriesInterpolateData:
 
     @td.skip_if_no_scipy
     def test_spline_interpolation(self):
-        # Explicit cast to float to avoid implicit cast when setting np.nan
-        s = Series(np.arange(10) ** 2, dtype="float")
+        s = Series(np.arange(10) ** 2)
         s[np.random.randint(0, 9, 3)] = np.nan
         result1 = s.interpolate(method="spline", order=1)
         expected1 = s.interpolate(method="spline", order=1)
